requirements.txt
certifi==2019.9.11
cloudinary==1.18.1
dj-database-url==0.5.0
Django==2.2.6
django-crispy-forms==1.7.2
django-filter==2.1.0
django-heroku==0.3.1
djangorestframework==3.9.2
gunicorn==19.9.0
mock==3.0.5
Pillow==6.1.0
psycopg2==2.7.7
pytz==2018.9
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
whitenoise==4.1.3
drf-writable-nested==0.5.1